Simpler python packaging
The python application fully relies on the standard python packaging and
setuptools which allows it to be installed with pip in the container.
The manifest is simplified to automatically pickup any new files.

The requirements.txt file is empty in the example so that the latest dependencies
will be used and can easily try out new things. When you want to freeze the
dependencies the container build will remind you what needs to be put in the
requirements.txt file. It is recommended that you freeze the dependencies for
production to make rebuilding a container more reproducable.

The usage of pytest is simplified to let you clearly define desired complexity
and line length limits. You can still use flake8-ignore if you need to but
it is not encouraged or shown in this example.

Eventlet has been removed as the recommended default way to run in gunicorn due
to unresolved compatibility isses with Python > 3.6 and the monkey patching that is
done to the ssl standard library. The memory saved is not worth the debugging time.

Whitenoise is used in the updated flask app to improve serving static files.