Flask app updates
The flask app really works now so it can be used as a starting point for a real app.

Using the generic flask_app package name simplifies changes needed when making a new app.
You can change it if you want but be careful to fix all the instances of it.
It can be generic since we are only pip installing to our container venv, if we wanted
to publish to pypy or install multiple packages the package name would need to be unique.

Containers are immutable so there is no configuration in the code that changes
depending on the environement the container is running in. Configuration files and
environment variables are provided at runtime.

Whitenoise is used to efficiently serve static files. This should make it easier to integrate
Javascript apps with a Flask API.

There are no dependencies on shared custom libraries. There were only 2 significant changes
needed to remove tools.* dependencies. The database connection wrapper was replaced by
directly using psycopg2, and require_weblogin_authentication was a small function that is
easy to include in each app. This should make it easy to automate the container building.