Full testing with tox
pytest with plugins for coverage and flake8 were not working correctly
when running in the tox virtualenvs.

Having tox run flake8 and coverage directly works well now.

Adding  bandit and safety is also easy to do with tox and provides additioanl
security checks that we haven't been able to integrate before.