Use flask app's logger everywhere
Flask stores a logger reference in the application context, and inside
blueprints we can use Flask's current_app magic object to get a handle
to that logger. Since we hook into gunicorn's logger to combine all the
app's logging into one stream, this is what we want. We don't need the
complication of holding on to our own logger in the app module and
importing it.current_app only works in a request context, so if you ever
need to get ahold of the logger at boot time in a blueprint you'll need
to export a reference to it from somewhere else after all.